% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.cluster.R
\name{Data.cluster}
\alias{Data.cluster}
\title{Perform Hierarchical Clustering on Predicted Microbial Abundance Data and Visualize Dendrograms}
\usage{
Data.cluster(
  predicted_data,
  clust_method,
  font_size = 0.2,
  auto_cutree = FALSE,
  dend_title_size = 15,
  cut_height_dist = 0.2
)
}
\arguments{
\item{predicted_data}{The output data frame from the \code{\link[MicrobTiSDA]{Pred.data}}.}

\item{clust_method}{A string, the agglomeration method to be used. This argument should be one of "ward.D", "ward.D2", "single",
"complete", "average", "mcquitty", "median", "centroid". Detail see \code{\link[stats]{hclust}}.}

\item{font_size}{A numeric value specifying the font size for text labels in the dendrogram plots (default: \code{0.2}).}

\item{auto_cutree}{Logical; if \code{TRUE}, the function automatically determines the optimal number of clusters based on
silhouette width (default: \code{FALSE}).}

\item{dend_title_size}{A numeric value specifying the font size of the dendrogram plot title (default: \code{15}).}

\item{cut_height_dist}{A numeric value used to adjust the vertical distance of the cut-off line annotation in the dendrogram
plot (default: \code{0.2}).}
}
\value{
A list with two elements: \code{cluster_results} contains the hierarchical clustering objects for each group,
and \code{cluster_figures} contains the corresponding ggplot2 dendrogram figures annotated with cluster information.
}
\description{
The \code{Data.cluster} function performs hierarchical clustering on predicted OTU time-series data to group
OTUs with similar temporal patterns. It computes a correlation-based distance matrix and applies hierarchical
clustering using a specified method. Depending on the \code{auto_cutree} parameter, the function either automatically
determines the optimal number of clusters using silhouette analysis or prompts the user to manually specify a
cut-off line. The resulting dendrograms are then annotated and returned along with the clustering objects.
}
\details{
For each group in the input \code{predicted_data}, the function first extracts the predicted OTU data (excluding the
last column, which is assumed to contain time information) and computes a correlation matrix, which is converted
into a distance matrix via
\deqn{d_{\text{corr}}(x,y) = 1-\frac{{\sum_{i=1}^{n}(x_i-\bar{y})}}{{\sqrt{{\sum_{i=1}^{n}(x_i-\bar{x})^2}} \sqrt{{\sum_{i=1}^{n}(y_i-\bar{y})^2}}}}}
where \eqn{x} and \eqn{y} represent the two OTU time series being compared, \eqn{n} denotes the total number of time points, and
\eqn{\bar{x}} and \eqn{\bar{y}} denote the means of the respective time series.

Hierarchical clustering is performed on the above distance matrix using the method specified in \code{clust_method}. When \code{auto_cutree} is \code{TRUE},
the function iteratively cuts the dendrogram for cluster counts ranging from 2 to n-1 (where n is the
number of OTUs) and calculates the average silhouette width for each case, selecting the optimal number of
clusters that maximizes the silhouette width. The dendrogram is then annotated with the suggested optimal
cluster number. If \code{auto_cutree} is \code{FALSE}, the dendrogram is displayed and the user is prompted
to input a cut-off height, which is used to annotate the dendrogram with a dashed line and corresponding label.
In both scenarios, the function returns a list containing the hierarchical clustering results and the
corresponding ggplot2 dendrogram figures.
}
\examples{
\dontrun{
# Assuming predicted_data is a list of predicted OTU abundance data frames:
cluster_output <- Data.cluster(predicted_data, clust_method = 'complete',
                               font_size = 0.3, auto_cutree = TRUE,
                               dend_title_size = 15, cut_height_dist = 0.2)
}
}
\author{
Shijia Li
}
