% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{fujita.data}
\alias{fujita.data}
\title{Example dataset 2 - The in vitro aquatic microbiota dataset.}
\format{
A data frame with 28 rows (ASVs) and 880 columns (samples).
}
\usage{
fujita.data
}
\description{
A count table used as an example for MicrobTiSDA functions.
}
\examples{
data("fujita.data")
head("fujita.data")
}
\keyword{datasets}
