% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{fujita.meta}
\alias{fujita.meta}
\title{Example dataset 2 - sample metadata.}
\format{
A data frame with 880 rows (samples) and 8 columns (variables).
}
\usage{
fujita.meta
}
\description{
Metadata corresponding to the samples in counts of the in vitro aquatic microbiota dataset.
}
\examples{
data("fujita.meta")
head("fujita.meta")
}
\keyword{datasets}
