% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{fujita.taxa}
\alias{fujita.taxa}
\title{Example dataset 2 - taxonomic information.}
\format{
A data frame with 565 rows (ASVs) and 9 columns (taxonomy levels).
}
\usage{
fujita.taxa
}
\description{
Taxonomic classification for the ASVs in the in vitro aquatic microbiota dataset.
}
\examples{
data("fujita.taxa")
head("fujita.taxa")
}
\keyword{datasets}
