% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interact.vis.R
\name{Interact.vis}
\alias{Interact.vis}
\title{Visualize Microbial Interaction Networks}
\usage{
Interact.vis(
  Interact_data,
  count_data,
  metadata,
  Subject_ID,
  Interact_threshold,
  Taxa = NULL,
  label_font_size = 3,
  label_distance = 0.15,
  title_size = 5,
  legend_title_size = 3,
  legend_text_size = 3
)
}
\arguments{
\item{Interact_data}{A list containing the inferred species interaction results generated by \code{\link[MicrobTiSDA]{Spec.interact}}.}

\item{count_data}{A matrix or data frame of the transformed species abundance data.}

\item{metadata}{A data frame. Containing information about all samples, including at least the grouping of all samples as well as
individual information (\code{Group} and \code{ID}), the sampling \code{Time} point for each sample, and other relevant information.}

\item{Subject_ID}{A character string specifying the column name in metadata that identifies subjects.}

\item{Interact_threshold}{A numeric value specifying the minimum absolute interaction strength to include in the network.}

\item{Taxa}{A data frame providing taxonomic annotations for microbial species.}

\item{label_font_size}{A numeric value specifying the font size for species labels in the network (default: 3).}

\item{label_distance}{A numeric value controlling the distance of labels from nodes (default: 0.15).}

\item{title_size}{A numeric value specifying the font size of the plot title (default: 5).}

\item{legend_title_size}{A numeric value specifying the font size of the legend title (default: 3).}

\item{legend_text_size}{A numeric value specifying the font size of the legend text (default: 3).}
}
\value{
A list of ggplot2 objects, each representing an interaction network for a subject.
}
\description{
This function visualizes microbial species interactions based on the given inter-species interaction results by \code{\link[MicrobTiSDA]{Spec.interact}}.
}
\details{
The function processes microbial interaction matrices from \code{\link[MicrobTiSDA]{Spec.interact}} to construct species interaction networks.
It selects edges with absolute interaction strengths above \code{Interact_threshold}, assigns interaction types (positive or negative), and
applies a circular layout for visualization. The function supports onpional taxonomic annotations for improved interpretability.

The nodes represent microbial species, sized according to their log-median abundance across samples. Edge colors indicate interaction types:
blue for negative and orange for positive interactions. Edge widths are scaled by interaction strength. If taxonomic annotations are provided,
node labels reflect species-level taxonomy; otherwise, OTU/ASV IDs are used.
}
\author{
Shijia Li
}
