% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.fold.CV.R
\name{k.fold.CV}
\alias{k.fold.CV}
\title{Compare the Performance of MicrobTiSDA with Several Other Regression Models}
\usage{
k.fold.CV(
  Data_for_Reg,
  pre_processed_data,
  unique_values,
  poly_degree = 3,
  k = 10,
  box_plot = TRUE,
  sig_label = FALSE
)
}
\arguments{
\item{Data_for_Reg}{The data frame output from the \code{\link[MicrobTiSDA]{Design}}.}

\item{pre_processed_data}{The transformed data output from the \code{\link[MicrobTiSDA]{Data.trans}} function. A
pre-processed OTU data frame with sample IDs as row names and OTU IDs as column names.}

\item{unique_values}{An integer specifying the minimum number of unique values required in an OTU for spline fitting (default: 5).}

\item{poly_degree}{A numeric value indicating the degree of the polynomial used in the GLM model (default: 3).}

\item{k}{An integer specifying the number of folds for cross-validation (default: 10).}

\item{box_plot}{Logical; if \code{TRUE}, boxplots of the CV RMSE values are generated (default: \code{TRUE}).}

\item{sig_label}{Logical; if \code{TRUE} (and if \code{box_plot = TRUE}), statistical significance annotations are
added to the boxplots (default: \code{FALSE}).}
}
\value{
If \code{box_plot = TRUE}, a list with two elements: an RMSE table summarizing the mean RMSE of each method for each OTU,
and a list of ggplot2 boxplots of the CV RMSE values. If \code{box_plot = FALSE}, only the RMSE table is returned.
}
\description{
This function performs k-fold cross-validation on different regression models and compare their performance in
predicting microbial feature time series data using root mean square error (RMSE).
}
\details{
The function randomly splits the dataset into k parts based on the k-fold parameter specified by the user. For
each fold, the dataset is divided into training and testing subsets, making sure that the time range of the
test set falls within the time range of the training set. Four regression models are then applied for fitting
and prediction: (1) Polynomial regression fitting the relationship between time and microbial feature using the
\code{MasigPro} package, with stepwise selection to determine the best model. (2) Using the \code{MetaDprof}
package's ssanova method to fit a smoothing spline model. (3) A locally weighted regression model fitted using
the \code{MetaLonDA} package. (4) A model fitted by trying degrees of freedom from 1 to 5 and choosing the optimal
degree based on GCV, followed by prediction. For each model, the function calculates predictions on the test set
and computes the RMSE by comparing the predictions with the actural values. he RMSE values are then compared to
evaluate the accuracy of the different methods in predicting time-series microbial feature data.
}
\examples{
\dontrun{
# Assuming 'design_data' is obtained from the Design function and 'otu_data' contains the pre-processed OTU data:
cv_results <- k.fold.CV(Data_for_Reg = design_data,
                        pre_processed_data = otu_data,
                        unique_values = 5,
                        poly_degree = 3,
                        k = 10,
                        box_plot = TRUE,
                        sig_label = TRUE)

# To view the RMSE table:
cv_results[['RMSE table']]

# To display the boxplots:
cv_results[['Boxplot of CV RMSE']]
}

}
\author{
Shijia Li
}
