% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{OSLO.infant.data}
\alias{OSLO.infant.data}
\title{Example dataset 1 - The OSLO term infant gut microbiota dataset.}
\format{
A data frame with 476921 rows (OTUs) and 77 columns (samples).
}
\usage{
OSLO.infant.data
}
\description{
A count table used as an example for MicrobTiSDA functions.
}
\examples{
data("OSLO.infant.data")
head("OSLO.infant.data")
}
\keyword{datasets}
