% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{OSLO.infant.meta}
\alias{OSLO.infant.meta}
\title{Example dataset 1 - sample metadata.}
\format{
A data frame with 77 rows (samples) and 4 columns (variables).
}
\usage{
OSLO.infant.meta
}
\description{
Metadata corresponding to the samples in counts of the OSLO Term infant gut microbiota dataset.
}
\examples{
data("OSLO.infant.meta")
head("OSLO.infant.meta")
}
\keyword{datasets}
