% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{OSLO.infant.taxa}
\alias{OSLO.infant.taxa}
\title{Example dataset 1 - taxonomic information.}
\format{
A data frame with 476921 rows (OTUs) and 8 columns (taxonomy levels).
}
\usage{
OSLO.infant.taxa
}
\description{
Taxonomic classification for the OTUs in the OSLO Term infant gut microbiota dataset.
}
\examples{
data("OSLO.infant.taxa")
head("OSLO.infant.taxa")
}
\keyword{datasets}
