% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.methods.R
\name{plot.DataOppCorVis}
\alias{plot.DataOppCorVis}
\title{Plot Method for DataOppCorVis Objects}
\usage{
\method{plot}{DataOppCorVis}(x, group, feature, ...)
}
\arguments{
\item{x}{A \code{DataOppCorVis} object containing correlation curve plots
in the \code{curves_plot} list element.}

\item{group}{A character string specifying the group to plot. Must be provided.}

\item{feature}{A character string specifying the feature within the group to plot.
If omitted, all features within the specified group will be plotted.}

\item{...}{Additional arguments}
}
\description{
This function prints correlation curve plots stored in a \code{DataOppCorVis} object.
Users can specify a group and optionally a specific feature within that group to plot.
}
