% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.methods.R
\name{plot.DataRFClassifier}
\alias{plot.DataRFClassifier}
\title{Plot Method for DataRFClassifier Objects}
\usage{
\method{plot}{DataRFClassifier}(x, ...)
}
\arguments{
\item{x}{A \code{DataRFClassifier} object containing margin scores and
cross-validation curve plots.}

\item{...}{Additional arguments}
}
\description{
This function plots the margin scores and cross-validation curve of a
\code{DataRFClassifier} object.
}
