% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.methods.R
\name{plot.MicrobTiSDA.cluster}
\alias{plot.MicrobTiSDA.cluster}
\title{Plot Method for MicrobTiSDA Cluster Objects}
\usage{
\method{plot}{MicrobTiSDA.cluster}(x, groups = NULL, ...)
}
\arguments{
\item{x}{A \code{MicrobTiSDA.cluster} object containing cluster plots
in the \code{cluster_figures} list element.}

\item{groups}{A character vector specifying the names of groups to plot.
If \code{NULL}, all groups in \code{x$cluster_figures} will be plotted.}

\item{...}{Additional arguments (currently not used) for compatibility with
generic \code{plot} methods.}
}
\description{
This function prints cluster plots stored in a \code{MicrobTiSDA.cluster} object.
Users can specify which groups to plot; if \code{groups} is \code{NULL},
all available cluster plots will be displayed.
}
