% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.methods.R
\name{plot.MicrobTiSDA.clusterCut}
\alias{plot.MicrobTiSDA.clusterCut}
\title{Plot Method for MicrobTiSDA Cluster Cut Objects}
\usage{
\method{plot}{MicrobTiSDA.clusterCut}(x, groups = NULL, ...)
}
\arguments{
\item{x}{x A \code{MicrobTiSDA.clusterCut} object containing cluster plots
in the \code{cluster_figures} list element.}

\item{groups}{A character vector specifying the names of groups to plot.
If \code{NULL}, all groups in \code{x$cluster_figures} will be plotted.}

\item{...}{Additional arguments (currently not used) for compatibility with
generic \code{plot} methods.}
}
\value{
The input \code{MicrobTiSDA.clusterCut} object, return invisibly.
}
\description{
This function prints cluster plots stored in a \code{MicrobTiSDA.clusterCut} object.
Users can specify which groups to plot; if \code{groups} is \code{NULL},
all available cluster plots will be displayed.
}
