% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.methods.R
\name{plot.MicrobTiSDA.MSERvisual}
\alias{plot.MicrobTiSDA.MSERvisual}
\title{Plot Method for MicrobTiSDA MSER Visualization Objects}
\usage{
\method{plot}{MicrobTiSDA.MSERvisual}(x, groups = NULL, clusters = NULL, ...)
}
\arguments{
\item{x}{A \code{MicrobTiSDA.MSERvisual} object containing feature cluster plots
in the \code{plots} list element.}

\item{groups}{A character vector specifying the names of groups to plot.
If \code{NULL}, all groups in \code{x$plots} will be plotted.}

\item{clusters}{An integer vector specifying which clusters to plot within
each selected group. If \code{NULL}, all clusters are plotted.}

\item{...}{Additional arguments}
}
\value{
The input \code{MicrobTiSDA.MSERvisual} object, return invisibly.
}
\description{
This function prints feature cluster plots stored in a
\code{MicrobTiSDA.MSERvisual} object. Users can specify which groups and
which clusters within each group to plot. If \code{groups} or \code{clusters}
is \code{NULL}, all groups or clusters are plotted.
}
