% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.methods.R
\name{plot.MicrobTiSDA.visual}
\alias{plot.MicrobTiSDA.visual}
\title{Plot Method for MicrobTiSDA Visual Objects}
\usage{
\method{plot}{MicrobTiSDA.visual}(x, groups = NULL, clusters = NULL, ...)
}
\arguments{
\item{x}{A \code{MicrobTiSDA.visual} object containing feature cluster plots
in the \code{plots} list element.}

\item{groups}{A character vector specifying the names of groups to plot.
If \code{NULL}, all groups in \code{x$plots} will be plotted.}

\item{clusters}{An integer vector specifying which clusters to plot within
each selected group. If \code{NULL}, all clusters are plotted.}

\item{...}{Additional arguments (currently not used) for compatibility with
generic \code{plot} methods.}
}
\description{
This function prints feature cluster plots stored in a \code{MicrobTiSDA.visual} object.
Users can specify which groups and which clusters within each group to plot.
If \code{groups} is \code{NULL}, all groups will be displayed. If \code{clusters}
is \code{NULL}, all clusters within the selected groups will be plotted.
}
