% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.methods.R
\name{plot.microbTiSDA_dynamic_vis}
\alias{plot.microbTiSDA_dynamic_vis}
\title{Plot Method for microbTiSDA Dynamic Visualization Objects}
\usage{
\method{plot}{microbTiSDA_dynamic_vis}(x, groups = NULL, ...)
}
\arguments{
\item{x}{A \code{microbTiSDA_dynamic_vis} object containing plots in the
\code{visualization} list element.}

\item{groups}{A character vector specifying the names of groups to plot.
if \code{NULL}, all groups in \code{x$visualization} will be plotted.}

\item{...}{Additional arguments}
}
\description{
This function iterates over all visualization elements stored in a
\code{microbTiSDA_dynamic_vis} object and prints each plot with its name.
}
