% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.methods.R
\name{plot.RfBiomarker}
\alias{plot.RfBiomarker}
\title{Plot Method for RfBiomarker Objects}
\usage{
\method{plot}{RfBiomarker}(x, ...)
}
\arguments{
\item{x}{A \code{RfBiomarker} object containing selected important OTUs,
predictions on training and test sets, and a cross-validation figure.}

\item{...}{Additional arguments}
}
\value{
The input \code{RfBiomarker} object, return invisibly.
}
\description{
This function prints the cross-validation figure of a \code{RfBiomarker} object
and displays basic information about the object contents.
}
