% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.codes.R
\name{Pred.data.MESR}
\alias{Pred.data.MESR}
\title{Predict Time-Series Data from Fitted Mixed-Effect Spline regression Models}
\usage{
Pred.data.MESR(Fitted_models, metadata, Group, time_step, Sample_Time)
}
\arguments{
\item{Fitted_models}{A list object generated by \code{\link[MicrobTiSDA]{Reg.MESR}}.}

\item{metadata}{A data frame. Containing information about all samples, including at least the grouping of all samples as well as
individual information (\code{Group} and \code{ID}), the sampling \code{Time} point for each sample, and other relevant information.}

\item{Group}{A character string indicating the column name in \code{metadata} that defines group information for prediction.}

\item{time_step}{A numeric value specifying the interval between new time points in the prediction sequence.}

\item{Sample_Time}{A character string indicating the column name in \code{metadata} that contains sample time information.}
}
\value{
An object of \code{PredictedDataMESR}. Each element of the list corresponds to a group and contains the predicted OTU abundances at the
new time points, with rows labeled by the group and time (formatted as \code{Group_T_Time}) and an additional column \code{Predicted_Time}.
}
\description{
This function generates predicted OTU abundances at new time points using previously fitted spline regression models.
For each group defined in the fitted models, it extracts the time range from the metadata, creates a new sequence of time points
with a specified interval, and uses the corresponding fitted models to predict values at these new time points.
}
\details{
The function accepts a list of fitted models (typically the output from \code{\link[MicrobTiSDA]{Reg.MESR}}) and sample metadata.
For each group (as defined by the \code{Group} parameter), it extracts the subset of metadata corresponding to that group. The minimum and maximum time
values are determined, and a new sequence of time points is generated using the provided \code{time_step}. For each OTU model of the
group, predictions are obtained via the \code{predict} function applied to the new time points, and the results are compiled into a data
frame with an additional column \code{Predicted_Time} indicating the prediction time. The output is a list of data frames, each corresponding
to a group, containing the predicted OTU abundances.
}
\examples{
\dontrun{
# Assuming Fitted_models is obtained from a spline regression
# fitting function (e.g., Reg.SPLR) and metadata is available:
predictions <- Pred.data.MESR(Fitted_models,
                             metadata,
                             Group = "Group",
                             time_step = 1,
                             Sample_Time = "TimePoint")
# View predictions for a specific group:
head(predictions[["Group1"]])
}
}
\author{
Shijia Li
}
