% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{preterm.data}
\alias{preterm.data}
\title{Example dataset 3 - The preterm infant gut microbiota dataset.}
\format{
A data frame with 202 rows (ASVs) and 51 columns (samples).
}
\usage{
preterm.data
}
\description{
A count table used as an example for MicrobTiSDA functions.
}
\examples{
data("preterm.data")
head("preterm.data")
}
\keyword{datasets}
