% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{preterm.meta}
\alias{preterm.meta}
\title{Example dataset 3 - sample metadata.}
\format{
A data frame with 51 rows (samples) and 36 columns (variables).
}
\usage{
preterm.meta
}
\description{
Metadata corresponding to the samples in counts of the in vitro aquatic microbiota dataset.
}
\examples{
data("preterm.meta")
head("preterm.meta")
}
\keyword{datasets}
