% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_datasets.R
\docType{data}
\name{preterm.taxa}
\alias{preterm.taxa}
\title{Example dataset 3 - taxonomic information.}
\format{
A data frame with 923 rows (ASVs) and 7 columns (taxonomy levels).
}
\usage{
preterm.taxa
}
\description{
Taxonomic classification for the ASVs in the in vitro aquatic microbiota dataset.
}
\examples{
data("preterm.taxa")
head("preterm.taxa")
}
\keyword{datasets}
