% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.DataOppCorVis}
\alias{print.DataOppCorVis}
\title{Information of Microbial features with opposite temporal shapes to users selected feature}
\usage{
\method{print}{DataOppCorVis}(x, ...)
}
\arguments{
\item{x}{An object of class DataOppCorVis}

\item{...}{Additional arguments}
}
\description{
Information of Microbial features with opposite temporal shapes to users selected feature
}
