% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.DataRFClassifier}
\alias{print.DataRFClassifier}
\title{Information of the constructed Random Forest classification model}
\usage{
\method{print}{DataRFClassifier}(x, ...)
}
\arguments{
\item{x}{An object of class DataRFClassifier}

\item{...}{Additional arguments}
}
\value{
The input \code{DataRFClassifier} object, returned invisibly.
}
\description{
Information of the constructed Random Forest classification model
}
