% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.Design}
\alias{print.Design}
\title{Print information of the Design matrix}
\usage{
\method{print}{Design}(x, ...)
}
\arguments{
\item{x}{An object of class Design.}

\item{...}{Additional arguments}
}
\value{
The input \code{Design} object, returned invisibly.
}
\description{
Print information of the Design matrix
}
