% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.FilteredData}
\alias{print.FilteredData}
\title{Print Method for FilteredData Object}
\usage{
\method{print}{FilteredData}(x, ...)
}
\arguments{
\item{x}{A \code{FilteredData} object}

\item{...}{Additional arguments}
}
\description{
Print Method for FilteredData Object
}
