% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.MicrobTiSDA.cluster}
\alias{print.MicrobTiSDA.cluster}
\title{Print the information of fitted temporal profile clusters}
\usage{
\method{print}{MicrobTiSDA.cluster}(x, ...)
}
\arguments{
\item{x}{An object of class MicrobTiSDA.cluster}

\item{...}{Additional arguments}
}
\value{
The input \code{MicrobTiSDA.cluster} object, returned invisibly.
}
\description{
Print the information of fitted temporal profile clusters
}
