% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.MicrobTiSDA.clusterCut}
\alias{print.MicrobTiSDA.clusterCut}
\title{Print the plot information of user selected feature clustering figures}
\usage{
\method{print}{MicrobTiSDA.clusterCut}(x, ...)
}
\arguments{
\item{x}{An object of class MicrobTiSDA.clusterCut}

\item{...}{Additional arguments}
}
\description{
Print the plot information of user selected feature clustering figures
}
