% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.MicrobTiSDA.interpolate}
\alias{print.MicrobTiSDA.interpolate}
\title{Print MicrobTiSDA.interpolate object}
\usage{
\method{print}{MicrobTiSDA.interpolate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{MicrobTiSDA.interpolate}}

\item{...}{Additional arguments}
}
\description{
Print MicrobTiSDA.interpolate object
}
