% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.MicrobTiSDA.MSERvisual}
\alias{print.MicrobTiSDA.MSERvisual}
\title{Information of visualizations of feature temporal patterns fitted with mixed-effect spline regression models}
\usage{
\method{print}{MicrobTiSDA.MSERvisual}(x, ...)
}
\arguments{
\item{x}{An object of class MicrobTiSDA.MESRvisual}

\item{...}{Additional arguments}
}
\value{
The input \code{MicrobTiSDA.MSERvisual} object, returned invisibly.
}
\description{
Information of visualizations of feature temporal patterns fitted with mixed-effect spline regression models
}
