% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.microbTiSDA}
\alias{print.microbTiSDA}
\title{Print MicrobTiSDA objects}
\usage{
\method{print}{microbTiSDA}(x, ...)
}
\arguments{
\item{x}{An object of class \code{microbTiSDA}}

\item{...}{Additional arguments}
}
\value{
The input \code{microbTiSDA} object, returned invisibly.
}
\description{
Print MicrobTiSDA objects
}
