% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.MicrobTiSDA.visual}
\alias{print.MicrobTiSDA.visual}
\title{Print the information of those clustered microbial features' temporal patterns}
\usage{
\method{print}{MicrobTiSDA.visual}(x, ...)
}
\arguments{
\item{x}{An object of class MicrobTiSDA.visual}

\item{...}{Additional arguments}
}
\value{
The input \code{MicrobTiSDA.visual} object, returned invisibly.
}
\description{
Print the information of those clustered microbial features' temporal patterns
}
