% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.microbTiSDA_dynamic_vis}
\alias{print.microbTiSDA_dynamic_vis}
\title{Print information of species interaction dynamic visualizations}
\usage{
\method{print}{microbTiSDA_dynamic_vis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{microbTiSDA_dynamic_vis}}

\item{...}{Additional arguments}
}
\value{
The input \code{microbTiSDA_dynamic_vis} object, returned invisibly.
}
\description{
Print information of species interaction dynamic visualizations
}
