% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.MicrobTiSDA_spline_regression}
\alias{print.MicrobTiSDA_spline_regression}
\title{Print information of fitted natural spline regression models}
\usage{
\method{print}{MicrobTiSDA_spline_regression}(x, ...)
}
\arguments{
\item{x}{An object of class MicrobTiSDA_spline_regression}

\item{...}{Additional arguments}
}
\description{
Print information of fitted natural spline regression models
}
