% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.PredictedData}
\alias{print.PredictedData}
\title{Print the information of fitted regression model predicted data}
\usage{
\method{print}{PredictedData}(x, ...)
}
\arguments{
\item{x}{An object of class PredictedData}

\item{...}{Additional arguments}
}
\description{
Print the information of fitted regression model predicted data
}
