% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.PredictedDataMESR}
\alias{print.PredictedDataMESR}
\title{Print method for PredictedDataMESR objects}
\usage{
\method{print}{PredictedDataMESR}(x, ...)
}
\arguments{
\item{x}{An object of class PredictedDataMESR.}

\item{...}{Additional arguments.}
}
\value{
The input \code{PredictedDataMESR} object, returned invisibly.
}
\description{
Print method for PredictedDataMESR objects
}
