% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.RegMESR}
\alias{print.RegMESR}
\title{Information of the fitted mixed-effect spline regression models}
\usage{
\method{print}{RegMESR}(x, ...)
}
\arguments{
\item{x}{An object of class RegMESR}

\item{...}{Additional arguments}
}
\description{
Information of the fitted mixed-effect spline regression models
}
