% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.methods.R
\name{print.TransformedData}
\alias{print.TransformedData}
\title{Print Method for TransformedData Object}
\usage{
\method{print}{TransformedData}(x, ...)
}
\arguments{
\item{x}{A \code{TransformedData} object}

\item{...}{Additional arguments}
}
\description{
Print Method for TransformedData Object
}
