% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.Design}
\alias{summary.Design}
\title{Summary Method for Design Objects}
\usage{
\method{summary}{Design}(object, ...)
}
\arguments{
\item{object}{An object of class \code{Design}.}

\item{...}{Additional arguments}
}
\value{
A summary list of the designed matrix.
}
\description{
Summary Method for Design Objects
}
