% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.FilteredData}
\alias{summary.FilteredData}
\title{Summary Method for FilteredData Object}
\usage{
\method{summary}{FilteredData}(object, ...)
}
\arguments{
\item{object}{A \code{FilteredData} object.}

\item{...}{Additional arguments}
}
\value{
A summary list containing dimensions and filtering parameters
}
\description{
Summary Method for FilteredData Object
}
