% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.MicrobTiSDA.cluster}
\alias{summary.MicrobTiSDA.cluster}
\title{Summary Method for MicrobTiSDA Cluster Objects}
\usage{
\method{summary}{MicrobTiSDA.cluster}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MicrobTiSDA.cluster}, typically produced by clustering functions in the \pkg{MicrobTiSDA} pacjage.}

\item{...}{Additional arguments.}
}
\value{
A summary list of the clustered objects.
}
\description{
Provides a summary of clustering results for objecs of class \code{MicrobTiSDA.cluster}. It reports the number of OTUs for each group, or
indicates if group was excluded.
}
