% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.MicrobTiSDA.clusterCut}
\alias{summary.MicrobTiSDA.clusterCut}
\title{Summary Method for MicrobTiSDA ClusterCut OBjects}
\usage{
\method{summary}{MicrobTiSDA.clusterCut}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MicrobTiSDA.clusterCut}, typically produced by cluster cutting function in the \pkg{MicrobTiSDA} package.}

\item{...}{Additional arguments}
}
\description{
Provides a summary of cluster cutting results for objects of class \code{MicrobTiSDA.clusterCut}. It reports the number of clusters for
each group, or indicates if no valid clustering was performend.
}
