% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.MicrobTiSDA.interpolate}
\alias{summary.MicrobTiSDA.interpolate}
\title{Summary method for MicrobTiSDA.interpolate objects}
\usage{
\method{summary}{MicrobTiSDA.interpolate}(object, ...)
}
\arguments{
\item{object}{A \code{MicrobTiSDA.interpolate} object created by the package functions.}

\item{...}{Additional arguments}
}
\description{
Provides a structure summary of a \code{MicrobTiSDA.interpolate} object, including its parameters and results.
It prints object class, the information of interpolated microbial count data and updated metadata.
}
