% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.microbTiSDA}
\alias{summary.microbTiSDA}
\title{Summary method for MicrobTiSDA objects
Provides a structure summary of a \code{microbTiSDA} object, including its parameters and results. It prints object class,
parameters settings, number of result groups/items, and previews of tabular of list-based result elements.}
\usage{
\method{summary}{microbTiSDA}(object, max_preview = 5, ...)
}
\arguments{
\item{object}{A \code{microbTiSDA} object created by the package functions}

\item{max_preview}{Integer. Maximum number of rows and columns to preview for tabular results. Default is \code{5}}

\item{...}{Additional arguments}
}
\description{
Summary method for MicrobTiSDA objects
Provides a structure summary of a \code{microbTiSDA} object, including its parameters and results. It prints object class,
parameters settings, number of result groups/items, and previews of tabular of list-based result elements.
}
