% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.MicrobTiSDA.visual}
\alias{summary.MicrobTiSDA.visual}
\title{Summary Method for MicrobTiSDA Visual Objects}
\usage{
\method{summary}{MicrobTiSDA.visual}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MicrobTiSDA.visual}, typically produced by visualization functions in the \pkg{MicrobTiSDA} package.}

\item{...}{Additional arguments}
}
\value{
A list of ggplot2 objects.
}
\description{
Provides a summary of visualization results for objects of class \code{MicrobTiSDA.visual}. It reports, for each group, the number of cluster-speficif
visualization plot available.
}
