% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.microbTiSDA_dynamic_vis}
\alias{summary.microbTiSDA_dynamic_vis}
\title{Summary Method for MicrobTiSDA Dynamic Interaction Visualization Objects}
\usage{
\method{summary}{microbTiSDA_dynamic_vis}(object, ...)
}
\arguments{
\item{object}{An object of class \code{microbTiSDA_dynamic_vis}, typically created by visualization functions in the \pkg{MicrobTiSDA} package.}

\item{...}{Additional arguments.}
}
\description{
Provides a concise summary of the dynamic interaction visualization results stored in a 'microbTiSDA_dynamic_vis' object.
The summary includes the number of nodes and edges in each group.
}
