% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.MicrobTiSDA_spline_regression}
\alias{summary.MicrobTiSDA_spline_regression}
\title{Summary Method for MicrobTiSDA Spline Regression Objects}
\usage{
\method{summary}{MicrobTiSDA_spline_regression}(object, ...)
}
\arguments{
\item{object}{An object of class \code{MicrobTiSDA_spline_regression}, usually created by spline regression functions in the \pkg{MicrobTiSDA} package.}

\item{...}{Additional arguments.}
}
\value{
A summary list of the fitted natural spline regression models.
}
\description{
Provides a concise summary for objects of class \code{MicrobTiSDA_spline_regression}. Prints the names of independent
variables and shows an example of the first fitted microbial feature.
}
