% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.PredictedData}
\alias{summary.PredictedData}
\title{Summary Method for PredictedData Objects}
\usage{
\method{summary}{PredictedData}(object, ...)
}
\arguments{
\item{object}{A \code{PredictedData} object returned by \code{Pred.data}.}

\item{...}{Additional arguments.}
}
\description{
Provides a concise summary of a \code{PredictedData} object,
including number of groups, number of time points, number of features,
and the time range for each group.
}
