% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.PredictedDataMESR}
\alias{summary.PredictedDataMESR}
\title{Summary method for PredictedDataMESR objects}
\usage{
\method{summary}{PredictedDataMESR}(object, ...)
}
\arguments{
\item{object}{An object of class PredictedDataMESR.}

\item{...}{Additional arguments.}
}
\description{
Summary method for PredictedDataMESR objects
}
