% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.RegMESR}
\alias{summary.RegMESR}
\title{Summary method for RegMESR objects}
\usage{
\method{summary}{RegMESR}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"RegMESR"}.}

\item{...}{Additional arguments}
}
\description{
Provides a concise summary of the results from a Regularized Multivariate
Exponential Spline Regression (RegMESR) analysis, including the number of
groups fitted, the total number of fitted models, and the model parameters.
}
