% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.TransformedData}
\alias{summary.TransformedData}
\title{Summary Method for TransformedData Objects}
\usage{
\method{summary}{TransformedData}(object, ...)
}
\arguments{
\item{object}{A \code{TransformedData} object returned by \code{Data.trans()}.}

\item{...}{Additional arguments.}
}
\value{
A summary list of the transformed data.
}
\description{
Summary Method for TransformedData Objects
}
