% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.methods.R
\name{summary.RfBiomarker}
\alias{summary.RfBiomarker}
\title{Summary method for RfBiomarker objects}
\usage{
\method{summary}{RfBiomarker}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"RfBiomarker"}.}

\item{...}{Additional arguments}
}
\value{
Summary information of the constructed random forest classification model.
}
\description{
Provides a concise summary of the results from a Random Forest biomarker
analysis, including the number of selected OTUs, their names, prediction
results on training and test sets, and information about cross-validation.
}
